requirements.txt
fastapi
uvicorn
openai
google-generativeai
youtube-transcript-api
pydantic
